hrms: employeeSearchCriteria Ids as list of Long

+++ rainmaker/egov-hrms/src/main/java/org/egov/hrms/web/contract/EmployeeSearchCriteria.java
@@ -31,7 +31,7 @@ public class EmployeeSearchCriteria {
 	
 	public Long asOnDate;
 	
-	public List<Integer> ids;
+	public List<Long> ids;
 	
 	public List<String> employeestatuses;
 	
@@ -39,7 +39,7 @@ public class EmployeeSearchCriteria {
 	
 	public List<String> uuids;
 	
-	public List<Integer> positions;
+	public List<Long> positions;
 	
 	public Boolean isActive;
